requirement.txt

Django==4.1.7
django-cors-headers==3.13.0

django-rest-swagger==2.2.0
djangorestframework==3.14.0
Pillow==9.4.0
requests==2.28.2

simplejson==3.18.3
sqlparse==0.4.3

